Rename folder
[x] rename folder `Sign-in-out-nav` to `SignInOut` to stick with a uniform folder naming convention.
