+++ client/src/App.jsx
@@ -3,7 +3,7 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import axios from 'axios';
 import MainSnackbar from './components/Snackbar/MainSnackbar';
 import { Route, Redirect } from 'react-router-dom';
-import SignInOut from './components/Sign-in-out-nav/SignInOut.jsx';
+import SignInOut from './components/SignInOut/SignInOut.jsx';
 import Landing from './components/Landing/Landing.jsx';
 // import DebugRoutes from './components/Debug/DebugRoutes.jsx';
 import { StripeProvider } from 'react-stripe-elements';
